+++ js/main.js
@@ -1,4 +1,4 @@
-// import './gallery.js'
+import './gallery.js'
